Add renaming of DetailToolbar and disable customer name/customer ref and adding of items to shipped/delivered invoices
