+++ packages/common/src/ui/layout/tables/RemoteDataTable.tsx
@@ -66,7 +66,7 @@ export const RemoteDataTable = <T extends DomainObject>({
         flexDirection: 'column',
       }}
     >
-      <MuiTable sx={{ display: 'block', overflow: 'auto' }}>
+      <MuiTable sx={{ display: 'block' }}>
         <TableHead sx={{ flex: 1, display: 'flex', flexDirection: 'column' }}>
           <HeaderRow>
             {columns.map(column => (
@@ -74,7 +74,15 @@ export const RemoteDataTable = <T extends DomainObject>({
             ))}
           </HeaderRow>
         </TableHead>
-        <TableBody sx={{ flex: 1, display: 'flex', flexDirection: 'column' }}>
+        <TableBody
+          sx={{
+            flex: 1,
+            display: 'flex',
+            flexDirection: 'column',
+            maxHeight: 500,
+            overflow: 'auto',
+          }}
+        >
           {data.map((row, idx) => {
             return (
               <DataRow<T>
